requirements.txt
discord.py
asyncio
flask
python-dotenv
PyGithub
